add: switch from preload_index to refresh_index
This is required to ensure that refresh_cache_ent is called for the
entries in the index cache.  refresh_cache_ent must be called so that
ce_mark_uptodate is called.  If cache entries are not marked as up-to-date
then add_excludes will not copy the OID from the index's cache.

This is a problem because the re-calculated OID will not match what's
in the untracked cache and prep_exclude will call invalidate_gitignore
which in turn marks the untracked_cache_dir as invalid.