add: include directory flags used by untracked cache
The untracked cache always uses the DIR_HIDE_EMPTY_DIRECTORIES
and DIR_SHOW_OTHER_DIRECTORIES flags (see new_untracked_cache).
Use the same flags in 'git add' when calling fill_directory
to increase the odds that the untracked cache can be used.

This commit only includes the setting of the two dir flags.
The changes to start using the untracked cache in 'git add' will
be included in a subsequent patch.

Signed-off-by: William Baker <William.Baker@microsoft.com>