fsmonitor: don't fill bitmap with removed entries
fill_fsmonitor_bitmap was populating fsmonitor_dirty
with all entries in the cache, even those that have
been marked for removal.  This caused a problem when
subsequent git commands read the index, as the bit's
position in fsmonitor_dirty would no longer match the
index in the cache.  Additionally, the number of entries
in fsmonitor_dirty could exceed the number of entries in
the cache.  If the number of entries in fsmonitor_dirty
exceeded the number of entries in the cache it resulted
in a segfault when tweak_fsmonitor was called because
ewah_each_bit would go beyond the bounds of the cache.

Signed-off-by: William Baker <William.Baker@microsoft.com>